Victory
By the grace of God, all these augmentations have been made, and I can
safely say that it is ready for release unto the masses! By God's grace
I had time to further enhance, test, and document this little fun
project!

Be sure to read the readme, and if you are reading this, you are awesome,
and I hope you have a grand day!
- Carter Slade